+++ lib/instream/connection/config.ex
@@ -4,7 +4,7 @@ defmodule Instream.Connection.Config do
   """
 
   @compile_time_keys [:loggers]
-  @defaults [
+  @global_defaults [
     loggers: [{Instream.Log.DefaultLogger, :log, []}],
     port: 8086,
     scheme: "http",
@@ -14,9 +14,10 @@ defmodule Instream.Connection.Config do
   @doc """
   Retrieves the compile time part of the connection configuration.
   """
-  @spec compile_time(atom, module) :: Keyword.t()
-  def compile_time(otp_app, conn) do
-    @defaults
+  @spec compile_time(atom, module, Keyword.t()) :: Keyword.t()
+  def compile_time(otp_app, conn, defaults \\ []) do
+    @global_defaults
+    |> Keyword.merge(defaults)
     |> Keyword.merge(Application.get_env(otp_app, conn, []))
     |> Keyword.take(@compile_time_keys)
   end
@@ -24,12 +25,14 @@ defmodule Instream.Connection.Config do
   @doc """
   Retrieves the runtime connection configuration for `conn` in `otp_app`.
   """
-  @spec runtime(atom, module, nil | nonempty_list(term)) :: Keyword.t()
-  def runtime(otp_app, _, [:otp_app]), do: otp_app
+  @spec runtime(atom, module, nil | nonempty_list(term), Keyword.t()) :: Keyword.t()
+  def runtime(otp_app, conn, keys, defaults \\ [])
 
-  def runtime(otp_app, conn, keys) do
-    otp_app
-    |> Application.get_env(conn, [])
+  def runtime(otp_app, _, [:otp_app], _), do: otp_app
+
+  def runtime(otp_app, conn, keys, defaults) do
+    defaults
+    |> Keyword.merge(Application.get_env(otp_app, conn, []))
     |> maybe_fetch_deep(keys)
     |> maybe_fetch_system()
     |> maybe_use_default(keys)
@@ -64,7 +67,7 @@ defmodule Instream.Connection.Config do
   defp maybe_fetch_system({:system, var}), do: System.get_env(var)
   defp maybe_fetch_system(config), do: config
 
-  defp maybe_use_default(config, nil), do: Keyword.merge(@defaults, config)
-  defp maybe_use_default(nil, keys), do: get_in(@defaults, keys)
+  defp maybe_use_default(config, nil), do: Keyword.merge(@global_defaults, config)
+  defp maybe_use_default(nil, keys), do: get_in(@global_defaults, keys)
   defp maybe_use_default(config, _), do: config
 end