+++ test/instream/connection/config_test.exs
@@ -106,4 +106,23 @@ defmodule Instream.Connection.ConfigTest do
     System.delete_env("INSTREAM_TEST_ENV_PASSWORD")
     System.delete_env("INSTREAM_TEST_ENV_USERNAME")
   end
+
+  test "inline configuration defaults" do
+    conn = Module.concat([__MODULE__, DefaultConfig])
+    key = :inline_config_key
+
+    Application.put_env(:instream, conn, [])
+
+    defmodule conn do
+      use Instream.Connection,
+        otp_app: :instream,
+        config: [{:"#{key}", "inline value"}]
+    end
+
+    assert "inline value" == conn.config([key])
+
+    Application.put_env(:instream, conn, Keyword.put(conn.config(), key, "runtime value"))
+
+    assert "runtime value" == conn.config([key])
+  end
 end